backend/pyproject.toml
[tool.poetry]
name = "tradeforge-aaas"
version = "1.0.0"
description = "Algorithm-as-a-Service Trading Platform with DeFi Integration"
authors = ["Ary HH <cateryatechnology@proton.me>"]
readme = "README.md"
license = "MIT"
repository = "https://github.com/cateryatechnology/tradeforge-aaas"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.109.0"
uvicorn = {extras = ["standard"], version = "^0.27.0"}
sqlalchemy = "^2.0.25"
alembic = "^1.13.1"
psycopg2-binary = "^2.9.9"
redis = "^5.0.1"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
pydantic = "^2.5.3"
pydantic-settings = "^2.1.0"
web3 = "^6.15.0"
ccxt = "^4.2.25"
pandas = "^2.1.4"
numpy = "^1.26.3"
streamlit = "^1.30.0"
plotly = "^5.18.0"
requests = "^2.31.0"
python-dotenv = "^1.0.0"
cryptography = "^42.0.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.4"
pytest-asyncio = "^0.23.3"
pytest-cov = "^4.1.0"
black = "^23.12.1"
flake8 = "^7.0.0"
mypy = "^1.8.0"
isort = "^5.13.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 100
target-version = ['py311']

[tool.isort]
profile = "black"
line_length = 100

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
